pyproject.toml
[tool.poetry]
name = "microblog"
version = "0.1.0"
description = "Microblog with FastAPI"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "^0.13.4"
fastapi = "^0.63.0"
SQLAlchemy = "1.3.23"
databases = "^0.4.2"
python-dotenv = "^0.15.0"
aiosqlite = "^0.17.0"
alembic = "^1.5.8"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
ipython = "^7.21.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
